requirements.txt
Django>=1.10
djangorestframework
django-filter
psycopg2
python-ldap
django-extensions
djangorestframework-jwt
esipy
eveapi
requests
django-rest-swagger